V1.1.5 submission to appstore (fantasy game types)

+++ IPAddFriendsViewController.m
@@ -244,7 +244,7 @@
                      NSLog(@"Request ID: %@", requestID);
                      self.facebookButton.enabled = NO;
                      RKObjectManager *objectManager = [RKObjectManager sharedManager];
-                     NSString *path = [NSString stringWithFormat:@"pool/addusers?pool_id=%d", self.poolID];
+                     NSString *path = [NSString stringWithFormat:@"pool/addusers?pool_id=%ld", (long)self.poolID];
                      [objectManager postObject:self.addUsers path:path parameters:nil success:^(RKObjectRequestOperation *operation, RKMappingResult *result) {
                          NSDictionary *dictionary = [NSDictionary dictionaryWithObjectsAndKeys:@"AddFBFriends",
                                                      @"type", @"Success", @"result", nil];
